requirements.txt
aiohttp[speedups]
aiomysql
aiofiles
ujson
git+https://github.com/Rapptz/discord.py
discord-ext-fslash
jishaku